modified how to calculate status and reason code count
no matter what's the reason code, status is always recorded. This is to make sure these data are in line with real time data. To avoid mutual access to database, the status will only increase, but not decrease. When the reason code button is pressed, the reason code count will increase in the database.

reading from sqlite and processing --- 55.41604208946228 seconds ---
writing to mongodb --- 0.3665323257446289 seconds ---